refactor: move nav to its own file
avoids the Next.js RSC boundary warning
